Allow setting ordering method at instantiation of MKLPardisoSolver object.
The fill-in reducing ordering calculation is always done during the
instantiation of the MKLPardisoSolver class.  However, the iparm
values cannot be edited until after the class is instantiated.
This means the default ordering method was always run when
MKLPardisoSolver was instantiated without a way of changing it.
This change fixes that.