Added factor method to MKLPardisoSolver class.
Previously, if `factor=False` was passed to the MKLPardisoSolver
instantiation, then the only way to factor the matrix was to call
`refactor(A)`. While this worked, `refactor` has unncessary
overhead as it resets the matrix data in MKL Pardiso. This new
`factor` method removes that overhead and factors the current
matrix without setting new data.